+++ app/services/history_recorder.rb
@@ -3,11 +3,12 @@ class HistoryRecorder
     @user = user
   end
 
-  def call(changes:, reference:, activity:)
+  def call(changes:, reference:, activity:, report: nil)
     changes.each do |value_changed, (previous_value, new_value)|
       HistoricalEvent.create(
         user: user,
         activity: activity,
+        report: report,
         reference: reference,
         value_changed: value_changed,
         previous_value: previous_value,